List all rentals with handlebars

+++ lions_share/db/migrate/20150717021408_create_listings.rb
@@ -2,9 +2,9 @@ class CreateListings < ActiveRecord::Migration
   def change
     create_table :listings do |t|
       t.string :type_of
-      t.decimal :price
-      t.decimal :bed_count
-      t.decimal :bath_count
+      t.integer :price
+      t.integer :bed_count
+      t.integer :bath_count
       t.string :neighborhood
       t.boolean :featured
       t.string :address
